cpuset: Simply allocation and freeing of cpumasks
Berbascum 2025, adapted to 4.14.314 for xiaomi vayu

Source: https://lkml.org/lkml/2018/10/12/870

The previous commit introduces a new subparts_cpus mask into the cpuset
data structure and a new tmpmasks structure.  Managing the allocation
and freeing of those cpumasks is becoming more complex.

So a number of helper functions are added to simplify and streamline
the management of those cpumasks. To make it simple, all the cpumasks
are now pre-cleared on allocation.