cpuset: Make CPU hotplug work with partition
Berbascum 2025, adapted to 4.14.314 for xiaomi vayu

Source: https://lkml.org/lkml/2018/10/12/868

When there is a cpu hotplug event (CPU online or offline), the partitions
may need to be reconfigured and regenerated. So code is added to the
hotplug functions to make them work with new subparts_cpus mask to
compute the right effective_cpus for each of the affected cpusets.
It may also change the state of a partition root from real one to an
erroneous one or vice versa.