BACKPORT: cpuset: Fix conflict ANDROID-cpuset-Make-cpusets-restore-on-hotplug
After the cpuset-Simply-allocation-and-freeing-of-cpumasks patch
is applied, next warning appears:

        /buildd/sources/kernel/cgroup/cpuset.c:546:1: warning: non-void function does not return a value in all control paths [-Wreturn-type]
}

The problem seems to be caused by an incompatibility
introduced by the Android patch:
cpuset-Make-cpusets-restore-on-hotplug

The cpuset-Simply-allocation-and-freeing-of-cpumasks patch
introduces some changes in the logic that breaks the
Android patch logic.

The problem seems to be a missing 'return NULL;'
in the first 'free_allowed:' label, in patched cpuset.c
after applying the backport patch.

This fix adds the missing return.

Signed-off-by: Berbascum <berbascum@ticv.cat>